refactor(bot/autoplay): extract markAsAutoplayTrack to queueMarkers.ts
Move markAsAutoplayTrack from queueEditOps.ts (via barrel export in
queueManipulation) to a new neutral module queueMarkers.ts. This breaks
Cycle A: queueManipulation → candidateCollector → autoplayAudit →
diversitySelector → markAsAutoplayTrack back to queueManipulation.

Since queueMarkers.ts doesn't import from queueManipulation or any
cycle member, the back-edge is severed.

- Created packages/bot/src/utils/music/autoplay/queueMarkers.ts
- Updated diversitySelector.ts to import from queueMarkers directly
- queueEditOps.ts re-exports from queueMarkers for backward compatibility
