+++ packages/bot/src/utils/music/autoplay/diversitySelector.ts
@@ -3,7 +3,7 @@ import { debugLog } from '@lucky/shared/utils'
 import { trackHistoryService } from '@lucky/shared/services'
 import { extractSongCore, cleanTitle, cleanAuthor } from '../searchQueryCleaner'
 import { calculateStringSimilarity } from '../duplicateDetection/similarityChecker'
-import { markAsAutoplayTrack } from '../queueManipulation'
+import { markAsAutoplayTrack } from './queueMarkers'
 import { extractYouTubeVideoId } from './scoringUtils'
 import type { RecommendationBasis } from './recommendationBasis.js'
 import { serializeBasis } from './recommendationBasis.js'
