+++ packages/bot/src/utils/music/autoplay/queueMarkers.ts
@@ -1,0 +1,49 @@
+import type { Track } from 'discord-player'
+
+/**
+ * Mark a track as autoplay-generated with optional metadata.
+ * Handles both mutable and sealed discord-player metadata objects.
+ */
+export function markAsAutoplayTrack(
+    track: Track,
+    recommendationReason: string,
+    requestedById?: string,
+): void {
+    const descriptor = Object.getOwnPropertyDescriptor(track, 'metadata')
+
+    if (descriptor?.configurable === false) {
+        // discord-player seals metadata as a non-configurable property on some
+        // track objects — Object.defineProperty would throw `Cannot redefine`.
+        // Mutate the object the getter/value returns directly (stable reference).
+        const meta = (
+            track as unknown as { metadata?: Record<string, unknown> }
+        ).metadata
+        if (meta && typeof meta === 'object' && !Object.isFrozen(meta)) {
+            meta['isAutoplay'] = true
+            meta['recommendationReason'] = recommendationReason
+            if (requestedById !== undefined)
+                meta['requestedById'] = requestedById
+        }
+        return
+    }
+
+    const existingMetadata =
+        (track as unknown as { metadata?: Record<string, unknown> }).metadata ??
+        {}
+    const existingRequestedById =
+        typeof existingMetadata.requestedById === 'string'
+            ? existingMetadata.requestedById
+            : undefined
+
+    Object.defineProperty(track, 'metadata', {
+        value: {
+            ...existingMetadata,
+            isAutoplay: true,
+            recommendationReason,
+            requestedById: requestedById ?? existingRequestedById,
+        },
+        writable: true,
+        configurable: true,
+        enumerable: true,
+    })
+}
