+++ packages/bot/src/utils/music/queueEditOps.ts
@@ -2,6 +2,8 @@ import { type Track, type GuildQueue } from 'discord-player'
 import { randomInt } from 'node:crypto'
 import { debugLog, errorLog } from '@lucky/shared/utils'
 import { replenishQueue } from './autoplay/replenisher'
+import { markAsAutoplayTrack } from './autoplay/queueMarkers'
+
 
 function randomIndex(maxExclusive: number): number {
     if (maxExclusive <= 1) return 0
@@ -166,49 +168,7 @@ export function extractSpotifyTrackId(track: Track): string | null {
     return match?.[1] ?? null
 }
 
-export function markAsAutoplayTrack(
-    track: Track,
-    recommendationReason: string,
-    requestedById?: string,
-): void {
-    const descriptor = Object.getOwnPropertyDescriptor(track, 'metadata')
-
-    if (descriptor?.configurable === false) {
-        // discord-player seals metadata as a non-configurable property on some
-        // track objects — Object.defineProperty would throw `Cannot redefine`.
-        // Mutate the object the getter/value returns directly (stable reference).
-        const meta = (
-            track as unknown as { metadata?: Record<string, unknown> }
-        ).metadata
-        if (meta && typeof meta === 'object' && !Object.isFrozen(meta)) {
-            meta['isAutoplay'] = true
-            meta['recommendationReason'] = recommendationReason
-            if (requestedById !== undefined)
-                meta['requestedById'] = requestedById
-        }
-        return
-    }
-
-    const existingMetadata =
-        (track as unknown as { metadata?: Record<string, unknown> }).metadata ??
-        {}
-    const existingRequestedById =
-        typeof existingMetadata.requestedById === 'string'
-            ? existingMetadata.requestedById
-            : undefined
-
-    Object.defineProperty(track, 'metadata', {
-        value: {
-            ...existingMetadata,
-            isAutoplay: true,
-            recommendationReason,
-            requestedById: requestedById ?? existingRequestedById,
-        },
-        writable: true,
-        configurable: true,
-        enumerable: true,
-    })
-}
+export { markAsAutoplayTrack}
 
 export function moveUserTrackToPriority(queue: GuildQueue, track: Track): void {
     const tracks = queue.tracks.toArray()
